First draft to upgrade to SpeziAccount 2.0
