+++ Sources/SpeziFirebaseAccount/Account Services/ReauthenticationOperationResult.swift
@@ -1,0 +1,13 @@
+//
+// This source file is part of the Stanford Spezi open-source project
+//
+// SPDX-FileCopyrightText: 2023 Stanford University and the project authors (see CONTRIBUTORS.md)
+//
+// SPDX-License-Identifier: MIT
+//
+
+
+enum ReauthenticationOperationResult {
+    case cancelled
+    case success
+}
