+++ Sources/SpeziFirebaseAccount/Views/FirebaseSignInWithAppleButton.swift
@@ -12,7 +12,8 @@ import SwiftUI
 
 
 struct FirebaseSignInWithAppleButton: View {
-    private let accountService: FirebaseIdentityProviderAccountService
+    @Environment(FirebaseAccountConfiguration.self)
+    private var service
 
     @Environment(\.colorScheme)
     private var colorScheme
@@ -22,12 +23,12 @@ struct FirebaseSignInWithAppleButton: View {
     @State private var viewState: ViewState = .idle
 
     var body: some View {
-        SignInWithAppleButton(onRequest: { request in
-            accountService.onAppleSignInRequest(request: request)
-        }, onCompletion: { result in
+        SignInWithAppleButton { request in
+            service.onAppleSignInRequest(request: request)
+        } onCompletion: { result in
             Task {
                 do {
-                    try await accountService.onAppleSignInCompletion(result: result)
+                    try await service.onAppleSignInCompletion(result: result)
                 } catch {
                     if let localizedError = error as? LocalizedError {
                         viewState = .error(localizedError)
@@ -39,14 +40,11 @@ struct FirebaseSignInWithAppleButton: View {
                     }
                 }
             }
-        })
+        }
             .frame(height: 55)
             .signInWithAppleButtonStyle(colorScheme == .light ? .black : .white)
             .viewStateAlert(state: $viewState)
     }
 
-
-    init(service: FirebaseIdentityProviderAccountService) {
-        self.accountService = service
-    }
+    nonisolated init() {}
 }
